pyproject.toml
[tool.poetry]
name = "dog-spotter-ai"
version = "0.1.0"
description = "🐩🐩🐩 犬を識別するAI！"
authors = ["osawa-koki <koki@osawa.dev>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
